# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "paopao-ce API",
    "description": "# paopao-ce API 文档\n\n欢迎使用 paopao-ce API 文档！\n\npaopao-ce 是一个开源的泡泡社区系统，提供完整的社交媒体功能。\n\n## 鉴权说明\n\n大部分API需要JWT Token进行鉴权，请在请求头中包含：\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## 响应格式\n\n所有API响应都遵循以下格式：\n```json\n{\n  \"code\": 0,\n  \"message\": \"success\",\n  \"data\": {}\n}\n```\n\n## 错误码说明\n\n- `0`: 成功\n- `10001`: 未授权\n- `10002`: 参数错误\n- `10003`: 权限不足\n- `10004`: 资源不存在\n\n## 分页参数\n\n支持分页的API使用以下查询参数：\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)\n",
    "version": "1.0.0",
    "contact": {
      "name": "paopao-ce",
      "url": "https://github.com/rocboss/paopao-ce"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8008",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.paopao.info",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/topic/follow": {
      "post": {
        "tags": ["社交关系"],
        "summary": "关注话题",
        "description": "关注指定话题，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token",
        "operationId": "followTopic",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowTopicReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "关注成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "话题不存在"
          }
        }
      }
    },
...
    "/user/follow": {
      "post": {
        "tags": ["社交关系"],
        "summary": "关注用户",
        "description": "关注指定用户，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 不能关注自己\n- 重复关注无效",
        "operationId": "followUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUserReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "关注成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/user/unfollow": {
      "post": {
        "tags": ["社交关系"],
        "summary": "取消关注用户",
        "description": "取消关注指定用户，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token",
        "operationId": "unfollowUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnfollowUserReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "取消关注成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/user/follows": {
      "get": {
        "tags": ["社交关系"],
        "summary": "获取关注列表",
        "description": "获取当前用户的关注列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "listFollows",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFollowsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/followings": {
      "get": {
        "tags": ["社交关系"],
        "summary": "获取粉丝列表",
        "description": "获取当前用户的粉丝列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "listFollowings",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFollowingsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          }
        }
      },
...
      "FollowTopicReq": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "话题ID"
          }
        }
      },
...
      "FollowUserReq": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          }
        }
      },
      "UnfollowUserReq": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          }
        }
      },
      "ListFollowsResp": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "total": {
            "type": "integer",
            "description": "总数"
          },
          "page": {
            "type": "integer",
            "description": "当前页码"
          },
          "page_size": {
            "type": "integer",
            "description": "每页数量"
          }
        }
      },
      "ListFollowingsResp": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "total": {
            "type": "integer",
            "description": "总数"
          },
          "page": {
            "type": "integer",
            "description": "当前页码"
          },
          "page_size": {
            "type": "integer",
            "description": "每页数量"
          }
        }
      },
...
    }
  }
}